+++ R/pat_scatterplot.R
@@ -4,7 +4,7 @@
 #' @title Compare all variables in a Purple Air Timeseries object
 #' 
 #' @description Creates a multi-panel scatterplot comparing all variables in the
-#' \code{pat} object. If any variables have no valid data, they are omitted from
+#' \emph{pat} object. If any variables have no valid data, they are omitted from
 #' the plot.
 #' 
 #' The list of available parameters include:
@@ -14,20 +14,21 @@
 #' \item{\code{pm25_A} -- A channel PM2.5 (ug/m3)}
 #' \item{\code{pm25_A} -- A channel PM2.5 (ug/m3)}
 #' \item{\code{temperature} -- temperature (F)}
-#' \item{\code{humidity} -- humidity (%)}
+#' \item{\code{humidity} -- humidity (\%)}
 #' \item{\code{uptime} -- seconds since last reset}
 #' \item{\code{acd0} -- analog input voltage}
 #' \item{\code{rssi} -- wifi signal strength (dBm)}
 #' }
 #' 
-#' @param pat Purple Air Timeseries "pat" object
+#' @param pat Purple Air Timeseries \emph{pat} object
 #' @param parameters vector of parameters to include
 #' 
-#' @return Tibble portion of the \code{pat} object, subset.
+#' @return Tibble portion of the \emph{pat} object, subset.
+#' 
 
 pat_scatterplot <- function(
   pat,
-  parameters = c('datetime','pm25_A','pm25_B','temperature','humidity')
+  parameters = c('datetime', 'pm25_A', 'pm25_B', 'temperature', 'humidity')
 ) {
   
   # For easier access
@@ -44,7 +45,9 @@ pat_scatterplot <- function(
   if ( sum(columnIsEmpty) > 0 ) {
     badColumns <- names(which(columnIsEmpty))
     badString <- paste(badColumns, collapse=", ")
-    message(paste("Columns: (", badString, ") have all missing data and are not shown."))
+    message(paste("Columns: (", 
+                  badString, 
+                  ") have all missing data and are not shown."))
   }
   
   plot(data, pch = '.')